Initial DTLTO commit
This follows on from our original RFC: here

This PR includes a design document which details the DTLTO design. There are some areas where we would like feedback from the upstream community to inform the design, those areas are noted in the design document.

This first commit just adds the initial scaffolding to show how DTLTO will work end-to-end. Support for additional features such as archive handling/caching etc.. will be added in later commits. Once this commit is in place we intend to add those features in a series of subsequent commits.

Questions to discuss with upstream contributors:

- What can we call the feature? DTLTO doesn't really distinguish this from Bazel-style DTLTO. Bazel-style DTLTO doesn't mean anything to anyone else!

- Perhaps there is a subset of arguments that should be automatically passed to the codegen tool from the driver level. For example perhaps the compiler should pass the relevant -mllm options to LLD if it sees: -fdiagnostics-format=msvc?

- Where to store temporary files? In particular we might want to discuss our downstream performance optimization (not included in this patch) for creating the native object files in the cache so that when we rename them to populate the cache we are guaranteed not to cross a device (which caused a measurable performance problem downstream).

- Our proposals for how to support bitcode archive members.

- Perhaps the (current) default tool should be clang -cc1 instead of just clang.