[DTLTO][LLVM] Allow LTO to take an AddBuffer function and use in DTLTO
Replace the use of AddStream in the DTLTO ThinLTO backend to add files
to the link with AddBuffer.

Unlike the InProcess ThinLTO backend, DTLTO runs the backend compilation
jobs by invoking an external process (currently clang). This writes the
output object file to disk. Therefore, DTLTO requires a performant way
of adding an existing file to the link. Note that the AddBuffer
mechanism is also used for adding a file to the link if there is a cache
hit.