[DTLTO][LLVM] clang format LTO.h to prevent automated checks errors
Some of the code in LTO.h did not conform to the LLVM coding standard.
However, it did match the style already used in that file.

However this was causing automated code-formatting checks from Github to
fail which was confusing on the PR. I decided to apply clang-format
everywhere to prevent this - even though the new code no longer matches
the style of the existing.