Remove dtlto-translate-options.ll and address some small mistakes
dtlto-translate-options.ll checks that the DTLTO code that creates Clang
command lines generates the expected options and that those options are
accepted by clang. I wrote it to use llvm-lto and opt so it could be
part of this commit before the clang and lld patches go in. However, I
think it is a nice principle to only use external tools in
cross-project-tests. I have removed the test from this PR with the
intention of adding it back later. Additionally if I can use the idea
of storing the Clang `-c` step options and applying them to the remote
backend compilations then this test will be rewritten to test that
behaviour instead.

Also address some minor mistakes that were identified by Teresa.