[LLD][COFF] Workaround for FASTBuild path normalisation issue
FASTBuild normalises the output file path when invoking Clang for the
backend compilations. This causes LTO importing to fail.

Unfortunately, LTO assumes that the path is exactly the same as that
supplied to the linker. The path is used as a lookup in various
data structures throughout the LTO code. It is also baked into various
hashes. A full fix for this issue will require some careful work.

For now, this commit works around the problem by normalising all
bitcode file paths to use Windows separators in COFF LLD - so that
FASTBuild's path normalisation should be a no-op.